add wrapers for DB and Filler class, prototype, doesn't work
