--- src/main/roboy/memory/Neo4j.java
@@ -1,34 +1,0 @@
-package memory;
-
-import org.neo4j.driver.v1.AuthTokens;
-import org.neo4j.driver.v1.Driver;
-import org.neo4j.driver.v1.GraphDatabase;
-
-import static memory.Config.*;
-
-public class Neo4j implements AutoCloseable {
-
-    private static Neo4j _instance;
-    private static Driver _driver;
-
-    private Neo4j() {
-        _driver = GraphDatabase.driver(NEO4J_ADDRESS, AuthTokens.basic(NEO4J_USERNAME, NEO4J_PASSWORD));
-    }
-
-    public static Driver getInstance() {
-        if (_instance == null) {
-            _instance = new Neo4j();
-        }
-
-        return _instance.getDriver();
-    }
-
-    private Driver getDriver() {
-        return _driver;
-    }
-
-    @Override
-    public void close() throws Exception {
-        _driver.close();
-    }
-}
