+++ src/main/roboy/memory/util/Config.java
@@ -1,0 +1,15 @@
+package memory.util;
+
+
+/**
+ * Configuration variables
+ */
+public class Config {
+    //ROS
+    public final static String ROS_MASTER_ADDRESS = "http://192.168.56.101:11311/";
+    //Neo4j
+    public final static String NEO4J_ADDRESS = "bolt://85.10.197.57:7687";
+    public final static String NEO4J_USERNAME = "memory";
+    public final static String NEO4J_PASSWORD = "*******"; //TODO: don't commit password to the repo
+
+}
